Part 4d full. tests users and authentication

+++ part4/blog-list/controllers/blogs.js
@@ -1,36 +1,62 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
+
 
 blogsRouter.get('/', async (req, res) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({})
+    .populate('user', { username: 1, name: 1 })
+
   res.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const { title, url, likes } = request.body
 
-  
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const user = request.user
+  const { title, author, url, likes } = request.body
   if (!title || !url) {
-    return response.status(400).json({
-      error: 'title and url are required'
-    })
+    return response.status(400).json({ error: 'title and url are required' })
   }
 
   const blog = new Blog({
-    title,
-    url,
-    likes: likes || 0,
-    author: request.body.author
+    title: request.body.title,
+    author: request.body.author,
+    url: request.body.url,
+    likes: request.body.likes || 0,
+    user: user._id
   })
 
   const savedBlog = await blog.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(404).end()
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(403).json({
+      error: 'only the creator can delete this blog'
+    })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
+
   response.status(204).end()
 })
+
+
 
 blogsRouter.put('/:id', async (request, response) => {
   const { title, author, url, likes } = request.body
